requirements.txt
asttokens==2.0.5
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
dataclasses==0.8
executing==0.8.2
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
icecream==2.1.1
idna==3.3
importlib-metadata==4.8.1
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
packaging==21.2
pluggy==1.0.0
py==1.10.0
Pygments==2.10.0
PyJWT==2.3.0
pyparsing==2.4.7
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.26
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
